refactor: split service and interfaces
TISNEW-5012
