+++ src/app/concern/concern.interfaces.ts
@@ -25,20 +25,3 @@ export enum ConcernStatus {
 
 // TODO: fill out types of concerns
 export enum ConcernType {}
-
-export const defaultConcern: IConcernSummary = {
-  concernId: null,
-  gmcNumber: null,
-  dateOfIncident: null,
-  concernType: null,
-  source: null,
-  dateReported: null,
-  employer: null,
-  site: null,
-  grade: null,
-  status: null,
-  admin: null,
-  followUpDate: null,
-  lastUpdatedDate: null,
-  comments: []
-};
