+++ src/app/concern/constants.ts
@@ -1,0 +1,18 @@
+import { IConcernSummary } from "./concern.interfaces";
+
+export const defaultConcern: IConcernSummary = {
+  concernId: null,
+  gmcNumber: null,
+  dateOfIncident: null,
+  concernType: null,
+  source: null,
+  dateReported: null,
+  employer: null,
+  site: null,
+  grade: null,
+  status: null,
+  admin: null,
+  followUpDate: null,
+  lastUpdatedDate: null,
+  comments: []
+};
