+++ src/app/concern/services/concern/concern.service.ts
@@ -1,8 +1,9 @@
 import { Injectable } from "@angular/core";
 import { HttpClient } from "@angular/common/http";
-import { IConcernHistory } from "../concern-history.interface";
+
 import { Observable } from "rxjs";
 import { environment } from "@environment";
+import { IConcernHistory } from "../../concern.interfaces";
 
 @Injectable({
   providedIn: "root"
@@ -14,9 +15,5 @@ export class ConcernService {
     return this.http.get<IConcernHistory>(
       `${environment.appUrls.getConcern}/${gmcNumber}`
     );
-  }
-
-  uploadFiles(formData: FormData) {
-    return this.http.post(environment.appUrls.uploadConcernDocuments, formData);
   }
 }
